MB: Check for the existence of a GN args location json.
Builders that define their GN args in starlark will be found via one of
the files specified in the gn_args_locations_files field in
mb_config.pyl. The config file includes both prod and dev versions, but
when a branch project is set up on a branch, the dev version is deleted
by the starlark generation since we don't run dev versions of the branch
projects. This causes MB operations that look up the config for a
builder to fail once a branch has a LUCI project set up. This changes it
so that it first checks for the existence of the location file before
attempting to read it.

Bug: b:303346412
Change-Id: Id6fa388713f21579b859aca6fb0ec7d11c0c2c10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4912198
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1205510}