[fedcm] Validate .well-known when it contains an accounts_endpoint.
In this CL, we expand the set of valid well-known files to include the
case where an accounts_endpoint is specified (as well as a signin_url)
and that matches the ones provided in the config file.

This allows multiple configURLs to be used, as long as they all share
the same accounts/signin endpoints, enabling IdPs to deploy more than
one endpoint configurations, while maintaining the privacy properties.

This functionality is developed behind a flag, while we get validation
from the privacy team that this isn't breaking any of the assumptions
we made so far.

Bug: 1429083
Change-Id: Ic7831357b9b7541d2038585b2781fb1b06a3a9b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4750824
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Sam Goto <goto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205512}