[a11y] Normalize the Textfield's (Views) accessible value
Textfields from Blink expose the display text (what's rendered on the
screen) as the accessible value, and we do the same for all labels both
in Blink or in Views.

However, the Views' Textfield class was the only class exposing the raw
text as accessible value. This was forcing us to manually handle the
case where it's a password and we should be exposing "***" instead,
and was completely ignoring the eliding logic.

This CL fixes it by applying the display text value in the RenderText,
the same way we are doing it for labels too.

This will facilitate my CL:4852850 where we are computing the rendered
offsets of each grapheme.

Change-Id: I6b1b80193f40be41939e86e209ec2bb6c9d45237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4912725
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1205580}