[views] Add missing metadata tag to more View subclasses.
Metadata is missing for a bunch of View subclasses. In an effort to start strictly enforcing Metadata - we need to backport existing classes to have the macro defined.

This is an ongoing effort with multiple CLs in an effort to split number of files.

This CL does have some small changes to UT that was using GetClassName to type check. See Patchset 4 for isolated changes.

Bug: 1485768
Change-Id: Id30d384dabddb51126a101fe546ff32dd8d3789f
AX-Relnotes: n/a.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4904624
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: David Yeung <dayeung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205590}