IndexedDB: make IndexedDBClientStateChecker non-associated
This mojo interface is used for checking BFCache state. Currently the
instance is initialized and passed around as an associated remote, which
means it will reuse the pipe of the mojo interface it's passed through,
which is the IndexedDBControl. There's no need to serialize calls
between these two interfaces.

Bug: 1378823
Change-Id: Ibf61f58f61117c05a3c3bb967d50599078daaa5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4904752
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205591}