LCPP: Add LcppStat verifier.
With the current implementation, there is no dedicated process to verify
the LcppData, especially LcppStat.  When the data is corrupted, the corrected
area is very limited. What to be verified are different between LearnLcpp
and Predict*.  We need to have unified way of verification.

Bug: 1419756
Change-Id: I5ac8f22ac6399e78f3eca16a5d24c8e1a627cad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905070
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205651}