Prerender: Attach TabHelpers for prerender-into-new-tab
This CL attaches TabHelpers to WebContents created for the
prerender-into-new-tab mode (also known as "prerendering WebContents",
see the design document below for details) by calling
TabHelpers::AttachTabHelpers().

This feature is behind the feature flag (kPrerender2InNewTab), so this
change shouldn't affect behavior in production.

> Details

The prerendering WebContents is not visible to users during prerendering
but becomes a proper browser tab on prerender activation. Therefore,
eventually it should have a full set of TabHelpers as a regular tab.
Instead of selectively attaching them to the WebContents during
prerendering, this CL attaches them all on the WebContents construction.
Some of the TabHelpers may not work during prerendering, but those cases
should be avoided by checking the lifecycle state. The CL author is now
auditing if the check is accordingly implemented in the TabHelpers:
https://docs.google.com/spreadsheets/d/1HjCoPLGgdCJbx4caKcsOeL3qIAFN0fk1wZqo3Nydqy4/edit?usp=sharing

> Design documents (Google internal):

- Overall design of the prerender-into-new-tab:
https://docs.google.com/document/d/1ZSwrsBtwgxO8xtgX0hA23YR92H62Y0o60ByVO352RL4/edit?usp=sharing

- WebContentsObserver, WebContentsDelegate, and TabHelpers on
  prerendering WebContents:
https://docs.google.com/document/d/1V0LPkS_sKcrOYNkpieP2N2SJGiWp4gNLizGsx8j5fgQ/edit?usp=sharing

Bug: 1350676
Change-Id: Ib54fabdc78f3553eed4f9718743ff4c43a3e660e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907409
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205695}