Add SurfaceFrameType::OVERLAP
As part of the shelf/caption insets separation project, we need to
place caption(header) in front of(z-axis) the window area instead of
putting it above(y-axis) the area.

This commit utilizes the existing wayland protocol `setFrame` and adds
a new enum SurfaceFrameType::OVERLAP to implement this feature. To
support SurfaceFrameType::OVERLAP, the change adjusts the way to
calculate window/client bounds, and to perform nonClientHitTest.

We also considered creating a new protocol for ARC to communicate the
position of caption. However, reusing the `setFrame()` function seems to
be a better solution.
For more details see go/arc-t-caption-migration-wayland.

Bug: b/282627319
Test: Manual testing to make sure it works when frame type is set to OVERLAY
Change-Id: Ie3584bfca510213cea66c4f0d32957eebd23c7a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4542005
Commit-Queue: Qijing Yao <yaoqq@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205709}