Give up trying to prevent copying HttpRequestHeaders
net::HttpRequestHeaders had a comment to try to make the type
uncopyable. In practice, copying HttpRequestHeaders objects is widely
done and frequently necessary. Remove the comment.

Remove the CopyFrom() method in favour of using the assignment operator
or copy constructor. In some places this is more efficient.

Change-Id: I381a80fb47b23347a785cda0ad6012861ded8b5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4904007
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Reviewed-by: Gang Wu <gangwu@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205710}