Support custom message in error and warn dialogs
Custom messages are now shown in error and warn dialogs, if defined.
To do this we introduce FilesPolicyDialog::Info (as a replacement
of FilesPolicyWarnSettings) and use it to store most of the dialog
related information such as the list of files, the message, and the
learn more link that should be shown to the user.

Bug: b/293122562
Change-Id: If8c5691b3f94268fe82c3c8170d1e2b0cd0feb15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866613
Reviewed-by: Aida Zolic <aidazolic@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Commit-Queue: Luca Accorsi <accorsi@google.com>
Cr-Commit-Position: refs/heads/main@{#1205718}