webview: Add feature flags and switches to crash for flag overrides
This should be helpful in knowing if a tester with a crash report had any flag overrides that could have contributed to their issue.

Verified this works:
http://crash/browse?q=reportid=%270b543683f17b0b76%27#2

The switches have their own section in the UI already:
https://screenshot.googleplex.com/5ujWQJwhBmgZ46Y

The chrome command line features show up in product data
(https://screenshot.googleplex.com/C7Qccderkuy2YUS)
because there is no crash parsing for those fields.
I've copied the format the ChromeOS folk follow.
We could always make the case in the future that this is worth while to
add a new section if needed.

This will only work for the browser process crashes since I am relying on the overrides fields and those aren't set in the render process.

Bug: 1484644
Change-Id: I4ba6a2da7ca03b6fe0be87f2f8765d5d67127ef2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4890218
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205745}