arc/tracing: model can generate no events in test
Accumulate all errors in ValidateGrahpicsEvents in the log rather than
bail on the first one. Also allow an empty generated list.  If there are
no events in it, the `ok` flag will be cleared when a required event is
not found, so there is no need to return early.

Test: make temporary change to generate an empty list and run the test
Bug: b/302382849
Change-Id: Iefe80d93b2d8297278f10821f2dc02d3d0fb27db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4911657
Reviewed-by: Yury Khmel <khmel@chromium.org>
Auto-Submit: Matthew DeVore <matvore@chromium.org>
Commit-Queue: Matthew DeVore <matvore@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205752}