Add an enum for autofill::Suggestion::icon
Right now icon is an std::string. We are planning on using this enum
instead of std::string everywhere in the code, which will be a very
lengthy CL, touching a lot of files.

This CL makes sure that the enum is in a final state before we start
using it everywhere else.

Bug: 1019660
Change-Id: I118373cdcca6388954091f16b8d26b14b3401178
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905256
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Theo Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1205764}