arc/tracing: get smaller trace JSON to avoid ooms
Before this CL we were subscribed to all viz events, which generated
roughly 50MB JSON as the raw trace results for 2 minute traces. This
could not be parsed on 4GB ChromeOS devices, and 20 minute traces could
not be parsed on 8GB devices. The JSON parser would run out of memory
and the Chrome process would crash.

In the trace config, register for only one category which generates only
a single event, which is the one we care about (vsync with new frame).
This allows 4GB devices to handle the JSON for 20 minute traces. The raw
JSON output of a 2 minute trace is reduced to about 160KiB.

Remove event matchers from the mapper for other viz events which we can
no longer receive and which arc-overview-tracing does not support.

Test: Trace an app window for 1200 seconds on a 4GB RAM nautilus
Bug: b/302382849
Change-Id: I2e5ed6a9782ad95cccc8940da725ecd5c65fca65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4909793
Auto-Submit: Matthew DeVore <matvore@chromium.org>
Commit-Queue: Matthew DeVore <matvore@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205768}