[Profiles DragDrop] Couple the drag object creation with initialisation
Having the initialisation separated from the creation of the drag
object is a source of uncertainty and a potential source of bugs (some
race conditions or double initialisation while waiting for the right
time to initialise the listeners).

Instead wait for the right time to completely create and initialise
the object properly and remove the notion of the object being in a
state where it is created but not usable.

Bug: 1466146
Change-Id: If6b23eddc99040f14719c3cfcc60f5431f4abba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4912553
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1205774}