dlp: Add is_dlp_managed field to DownloadItem
Adds a field and related accessors to mark a DownloadItem as managed by
DLP, i.e. that there is at least one DLP Files Restriction rule that
applies to that downloads, so the file might be protected and some of
the actions on it blocked/warned and/or reported to the admin.

R=qinmin@chromium.org

Fixed: b:302506453
Change-Id: I7a8288e7cdc36dd30b4c7237af94d7f44c140107
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4909956
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205797}