Reland "[PwaRestore]: Flesh out the second sheet a bit more."
(this time with build fix for Components UnitTests, see
one-liner change in patchset 2).

This change adds the Deselect All button at the bottom
(which calls a function that does nothing), as well as
an empty ScrollView and polishes the layout a bit to
look more like the mocks.

Bug: 1481570, 1489608
Change-Id: I4bf91551b7b28c6e1aad112cf239cba11718829b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4912482
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Glenn Hartmann <hartmanng@chromium.org>
Auto-Submit: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205805}