[Hub] Create HubManager
Create a HubManager to act as the external API to act as the interface
between Hub and the outside world.

HubLayout will need access to the View that hosts the Hub to attach it
to the view hierarchy and use it for animations. Landing this first
allows us to more easily build and test HubLayout.

Removed PaneManagerFactory in favor of HubManagerFactory.

Bug: 1487315
Change-Id: I51afb7c2f0e6a0274044d44479bc992a2236191a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4898977
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1205806}