Set plain text password in UserContext in relevant flows
This CL makes sure that all login-related flows have
the plain text password input passed in a way that
allows to distinguish the origin of the password (GAIA,
SAML, local input) further in the flow. Other flows
like unlock are not in the scope of this CL.

This is a part of longer initiative to eventually get rid
of "Key" type that key origin, which is becomes relevant
with introduction of local passwords.

Bug: b:291808449
Change-Id: I9f87ea92cdfbc23d7970f8785b29cfb94cefc93b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905200
Reviewed-by: Roland Bock <rbock@google.com>
Commit-Queue: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: Anastasiia N <anastasiian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205814}