Implement an unprefixed mask-image property
This patch adds an updated and unprefixed mask-image property behind
the experimental flag CSSMaskingInterop. A new test has been added for
cases with both mask-image and -webkit-mask-image, which is important
for web compatibility. Test expectations have been updated for the
newly-passing tests. In addition to unprefixing, the patch removes
support (also behind the flag) for the non-standard cross-fade
interpolation of mask-images, which now matches both WebKit and Gecko.

The implementation approach in css_properties.json5 is novel, and is
based on a prototype by andruud@chromium.org. The approach is to add
a new flag-guarded property, "mask-image", which is a copy of the
existing "-webkit-mask-image" property. Then, a new flag-guarded
alternative of "-webkit-mask-image" is added, which is an alias of the
"mask-image" property ("-webkit-alternative-mask-image"). This ensures
that "-webkit-mask-image" uses the new "mask-image" code when the flag
is enabled. To ensure the "-webkit-mask" shorthand property uses the new
"mask-image" property, a flag-guarded alternative of "-webkit-mask" is
also added.

To remove support for interpolation, the new "mask-image" property does
not have "interpolable: true".

Bug: 432153
Change-Id: Ia69a2174f09638701083570ba838fcc9f546186c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907773
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205827}