[iOS] Fix History and tabs toggle when both types are disabled by policy.
For kSignedIn users, “History and Tabs” toggle is disabled by policy
only when both History and Tabs are disabled.
This logic was added before to updateSyncItemsNotifyConsumer, but it
should also be added to tableViewItemWithDataType where the toggles are
first created.

This CL adds the logic to tableViewItemWithDataType, and adds
corresponding EG tests.

Fixed: 1487909
Change-Id: I0eb94a10036b33fe8883cc91b61f2430fcacf010
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4908262
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1205831}