[mathml] Tweak MathItalicCorrection and add unit test for a edge case.
Code coverage reports the case !harfbuzz_face as uncovered [1]. In
practice, MathItalicCorrection is indeed only used in
StretchyOperatorShaper::Shape where this condition is checked before.
This CL strengthen a bit this by actually checking whether the font
has a MATH table, aligning the behavior with the documentation in the
header file [2] and with the MathConstant method. A unit test is also
added to cover that case.

[1] https://analysis.chromium.org/coverage/p/chromium/file?path=//third_party/blink/renderer/platform/fonts/opentype/open_type_math_support.cc&revision=c8a01d35ed430e49e8275f641f735e689a97c105#131
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/fonts/opentype/open_type_math_support.h;l=91;drc=047c7dc4ee1ce908d7fea38ca063fa2f80f92c77

Bug: 1464154
Change-Id: I7f070376d7d111fcfd33b5c875d5268f7676556f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4689687
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1205856}