Merge source and trigger interop input lists
While this change retains the ordering of registrations in the JSON
files for ease of review, in the future this will allow us to sort
sources and triggers by timestamp, which will make tests easier to
understand: currently it is tedious to go back and forth between the
two lists to understand the ordering.

Additionally, this increases the testable surface by allowing detection
of registrations that attempt to register sources and triggers
simultaneously.

Change-Id: I2ba356727004a2fff8074cf92da4a4e39c397866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916351
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205934}