Add rest of the CPU performance intervention UMA metrics
This implements the advanced UMA from
go/performance-interventions-feasibility to build on the initial
metrics logged in a prior CL.

At a high level, when we detect CPU is over the threshold, we log
an "Immediate" version of the metrics and then wait a (Finchable)
delay before logging the "Delayed" version of the metrics if the
threshold is still exceeded. A separate UMA will track how long
the threshold was exceeded (assuming the browser drops below the
threshold at some point).

Bug: 1481051
Change-Id: I439bc0824c075bc2fcf2650a2104cd8963a1ebe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4888868
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Alison Gale <agale@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1205974}