Make GetPreviousRenderFrameHostId resilient to current RFH changes
NavigationRequest::GetPreviousRenderFrameHostId() currently returns
the ID of the current RFH at the time the NavigationRequest is
constructed. This means if another RenderFrameHost committed and
replace the current RFH while the navigation is still in progress,
the saved RFH id might become stale.

After auditing [1], all current usage of the function turns out to
want to refer to the latest value of the current RFH up until
the navigation commits and potentially swaps in a new RFH, so this
CL makes it so that the function returns the current RFH if the
navigation hadn't committed a new RFH, and a snapshotted ID of
the current RFH just before the potential swap happens afterwards.

We do still track the original current RFH from request
construction time in current_render_frame_host_id_at_construction_,
but it's currently only needed within NavigationRequest itself, so
it's not exposed outside.

Note: navigation queueing and the early RFH swap are two ways
to change the current RFH in the middle of a navigation.

[1]: https://docs.google.com/document/d/1FSnuLm9vuqjR89v2KD5CcjdAksDs-kZn_XAX4W8yrsU/edit?usp=sharing

Bug: 1488411
Change-Id: I6e10577f6d996fca674a8201aa98516b475911b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4904258
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1205990}