[ios] Fix crash when returning a pointer style for the NTP fakebox
This seems to be caused by a race condition where the user hovers over
the fakebox, but before iOS asks for a pointer style the view is removed
from the window. In that case, we no longer need to return a pointer
style.

Fixed: 1487697
Change-Id: I68fb5dcc11f3497db6c865af1c248caf72903ee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916617
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1206037}