[Read Aloud] Use language code to filter voices.
The "default" property on SpeechSynthesisVoice doesn't always
return an expected default voice. e.g. on Linux, a Dutch voice is
default, even though the device and browser language are English.

Still attempt to use the default voice from the list of voices in the
language or all the voices if no voices are found in the default
language.

More work will be done on voice / language management, but this
just enables faster development on Linux with an English voice
for now.

Bug: 1474951
Change-Id: I271c7a50476308520db65f182031306842576524
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4917492
Commit-Queue: Lauren Winston <lwinston@google.com>
Reviewed-by: Kristi Saney <kristislee@google.com>
Cr-Commit-Position: refs/heads/main@{#1206045}