Extract the drag policy code from the Aura drag-and-drop code and reuse it on the Mac
There's been quite a lot of hardening that was applied to the Aura
drag and drop code that wasn't applied to the Mac version. Extract
Aura's DragStart struct, and code using it, to reuse it for the Mac.
This fixes the issue by reusing the hardening written for the Aura
implementation, and hopefully ensures that in the future, changes
to the hardening will be applied to the Mac's drag and drop code
as well.

Fixed: 1488621
Low-Coverage-Reason: HARD_TO_TEST plus this is a security catchup
Change-Id: I2d6999de54168a3b2b71e017acbe86028bdc62a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4913675
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206046}