Prerender: Make sure prerendering WebContents has an initial empty document
As a follow-up for [1], this CL makes sure that WebContents created for
prerender-into-new-tab has an initial empty document as the primary
page. This condintion is important as WebContentsObservers attached to
the WebContents may assume that there is the primary page and access it
during prerendering.

[1]https://chromium-review.googlesource.com/c/chromium/src/+/4907409/comment/9ef853f9_5ad430df/

Bug: 1489542
Change-Id: I5694882d9e6f76130dd88b48031c12dfaa468595
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4914796
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206078}