Improve accuracy of JS execution error message.
The existing code will return an error of "Renderer terminated" even
when the actual error is that the underlying RenderFrameHost that was
the target of the JS execution has just been deleted, even though the
renderer process may not have exited. This CL improves the error message by making it specific to the case at hand.

See
https://chromium-review.googlesource.com/c/chromium/src/+/4856149/comment/10d16b7a_6c018e17/
for where this was discovered.

Bug: 1478463
Change-Id: I76505a0f923a15df11dd6cca18baf3a53b166e3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916913
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206085}