Fix RP overlay clipping, and disable RP Transform delegation
Predicting buffer size in the overlay factory is difficult, so
there were bugs when clipping and setting UVs. Use the
old way of baking the clip_rect into the display_rect, and
settings UVs in SkiaRenderer.

Also temporarily disable delegating RPs with matrix
Transforms because accounting for clipping and filters
that expand bounds is more complicated. I'm working on
fixing that in a follow up and re-enabling transform
delgation for RPs.

Bug: 1471182
Change-Id: I3c5f066bebd3a47982fc250fc49cb676c798bc34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907881
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Commit-Queue: Kevin Haslett <khaslett@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1206098}